product_service/requirements.txt
Django==1.6.11
mysqlclient==1.3.14
pymysql==0.9.3
cryptography==2.3.1
gunicorn==19.10.0
django-environ==0.4.5
django-debug-toolbar==1.3.2
requests~=2.27.1